required_data_for_p3_healthcare_RCM/requirements.txt
pandas==2.2.1
mysql-connector-python==8.3.0
google-cloud-bigquery==3.17.2
google-auth==2.29.0
google-cloud-core==2.4.1
sqlalchemy==2.0.30
pyarrow==16.1.0
graphviz
gcsfs
matplotlib
numpy==1.26.4
db-dtypes==1.2.0
openpyxl==3.1.2
PyMySQL==1.1.1
google-cloud-storage==2.18.2
db-dtypes==1.2.0
seaborn==0.13.2
python-dotenv==1.0.1
loguru==0.7.2
pytest==8.2.2
